fix restart req for locked addons
Locked addons needed the game to be restarted for the update to be applied.
This is no longer the case since the addon is now immediately unloaded instead, the update downloaded and then immediately loaded with the new version if available.

The parameters of the UpdateAddon func had to be adjust to accommodate this, the tmpDefs have to be cached once more for chain reloads.