Added the `/utf-8` flag to compiler options and remaining cleanup.
This flag tell the compiler to always interpret the input and produce output as utf8, the sensible solution.
The insane default is again; to use your system codepage (works on my machine :tm:).

this also enables character validation, which was the thing that caught the remaining cases of "incorrectly" encoded characters.